fix build_solution fix make_solution add helper stringInt32 add helper stringFloat fix stringalized Color rework tools Ravinggen improve ToolUtils

+++ src/Tools/AstralaxCompiler/AstralaxCompiler.cpp
@@ -200,7 +200,7 @@ int APIENTRY WinMain( HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdL
     }
 
     fseek( f, 0, SEEK_END );
-    int f_size = ftell( f );
+    long f_size = ftell( f );
     rewind( f );
 
     if( f_size == 0 )
