+++ src/Tools/TexturePacker/TexturePacker.cpp
@@ -278,7 +278,7 @@ int APIENTRY WinMain( HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdL
         }
 
         fseek( f, 0, SEEK_END );
-        int f_size = ftell( f );
+        long f_size = ftell( f );
         rewind( f );
 
         void * buff = malloc( f_size );
